+++ NRFM_Auditoria/NRFM_Auditoria/Form1.cs
@@ -8,131 +8,12 @@ namespace NRFM_Auditoria
     {
 
         // filtro utilizado para solo mostrar acrhivos de excel
-        public const string FILTROS_ARCHIVOS_EXCEL = "Archivos Excel|*.xlsx;*.xlsm;*.csv;";
-        public const string NOMBRE_COL_RESPONSABLE = "Responsable";
         public const string RUTA_ARCHIVOS_RESPONSABLES = "Extracciones";
 
-        public const int MAX_FILA_CABECERA = 1000;
         public Form1()
         {
             InitializeComponent();
         }
-
-        public string obtenerArchivoSeleccionado()
-        {
-            /*
-                Obtiene la ruta del archivo seleccionada en el explorador de archivos desplegado, si no se selecciona
-                un archivo retorna null
-            */
-            OpenFileDialog ofd = new OpenFileDialog();
-            ofd.Filter = FILTROS_ARCHIVOS_EXCEL;
-
-
-            if (ofd.ShowDialog() == DialogResult.OK)
-            {
-                return ofd.FileName;
-            }// si el archivo esta ok
-
-            return null;
-        }
-
-        public int encontrarCabecera(IXLWorksheet hoja)
-        {
-            /*
-                Busca en el archivo la primera celda de color negro que se encuentre en la primera columna para determinar que 
-                es el fin de la cabecera, o en su defecto el encabezado de los datos. Si no encuentra una cabecera en el numero maximo
-                de filas, entonces regresa -1.
-            */
-            int index = 1;
-
-            // hay que buscar el fin de cabecera
-            while (!hoja.Cell("A" + index.ToString()).Style.Fill.BackgroundColor.Equals(XLColor.FromTheme(XLThemeColor.Text1)) && MAX_FILA_CABECERA > index)
-            {
-                index++;
-            }//fin while encontrar cabecera
-
-            //si no se encuentra la cabecera en el limite, se omite la hoja
-            if (MAX_FILA_CABECERA <= index)
-            {
-                MessageBox.Show("No se ha encontrado la cabecera en " + hoja.Name);
-                return -1;
-            }// fin maximo de columnas cabecera
-            return index;
-        }
-
-        public char encontrarColResponsable(IXLWorksheet hoja, int filaEncabezado)
-        {
-            /*
-                Indicando la fila donde supuestamente se encuentra el encabezado de los datos (los titulos), recorre las columnas
-                hasta encontrar la celda cuyo valor sea igual a NOMBRE_COL_RESPONSABLE, devolviendo la letra de dicha columna. En 
-                caso de no encontrarla retorna "-".
-            */
-            // buscamos en donde se encuentra la columna de responsables
-            char colResponsable = 'A';
-
-            while (!hoja.Cell(colResponsable + filaEncabezado.ToString()).IsEmpty())
-            {
-                //Debug.WriteLine(hoja.Cell(colResponsable + index.ToString()).Value.ToString());
-                if (hoja.Cell(colResponsable + filaEncabezado.ToString()).Value.ToString() == NOMBRE_COL_RESPONSABLE)
-                {
-                    return colResponsable;
-                }//fin if celda con valor Responsable
-                colResponsable++;
-            }//fin while encontrar col responsable
-            return '-';
-        }
-
-        public int finColumnasArchivo(IXLWorksheet hoja, int filaEncabezado)
-        {
-            /*
-                Retorna la columna final del encabezado de los datos en el archivo
-            */
-            char columna = 'A';
-
-            while (!hoja.Cell(columna + filaEncabezado.ToString()).Style.Fill.BackgroundColor.Equals(XLColor.FromTheme(XLThemeColor.Text1)))
-            {
-                columna++;
-            }//fin while encontrar col responsable
-            return columna;
-        }
-        public bool archivoTieneHojas(XLWorkbook archivo)
-        {
-            /*
-                Recorre las hojas del archivo, a la primera retorna true.
-            */
-            foreach (IXLWorksheet hoja in archivo.Worksheets)
-            {
-                return true;
-            }// fin for recorre las hojas del archivo nuevo
-            return false;
-        }
-
-        public int encontrarFinDatos(IXLWorksheet hoja, int FinCabecera)
-        {
-            /*
-                Retorna el numero de la fila donde terminan los datos.
-            */
-
-            int index = FinCabecera;
-            while (!hoja.Cell("A" + index.ToString()).IsEmpty())
-            {
-                index++;
-            }
-            return index;
-        }
-
-        public string estandarizarNombre(string nombre)
-        {
-            /*
-                Elimina los espacios de mas en el nombre del responsable. Separa los nombres por espacio
-                y los reagrupa en una nueva cadena.
-            */
-            string[] Nombres = nombre.Split(new[] { " " }, StringSplitOptions.RemoveEmptyEntries);
-
-            string resultado = string.Join(" ", Nombres);
-            return resultado.ToUpper();
-        }
-
 
         private void cargarArchivo_Click(object sender, EventArgs e)
         {
@@ -147,10 +28,10 @@ namespace NRFM_Auditoria
             if (!Directory.Exists(ruta_carpeta))
             {
                 Directory.CreateDirectory(ruta_carpeta);
-            }// fi no existe carpeta Extraccion
+            }// fi no existe carpeta Extraccion 
 
             // se selcciona el archivo a leer
-            string ruta_archivo = obtenerArchivoSeleccionado();
+            string ruta_archivo = FuncionesAuditoria.obtenerArchivoSeleccionado();
 
             if (ruta_archivo != null)
             {
@@ -165,7 +46,7 @@ namespace NRFM_Auditoria
 
                 foreach (IXLWorksheet hoja in archivo.Worksheets)
                 {
-                    int Fin_Cabecera = encontrarCabecera(hoja);
+                    int Fin_Cabecera = FuncionesAuditoria.encontrarCabecera(hoja);
                     if (Fin_Cabecera == -1)
                     {
                         continue; // si no se encontro cabecera que pase a la siguiente hoja
@@ -173,10 +54,10 @@ namespace NRFM_Auditoria
                     else
                     {
                         // buscamos la columna de responsable
-                        char colResponsable = encontrarColResponsable(hoja, Fin_Cabecera);
+                        char colResponsable = FuncionesAuditoria.encontrarColResponsable(hoja, Fin_Cabecera);
                         if (colResponsable == '-')
                         {
-                            MessageBox.Show("No se ha encontrado la columna " + NOMBRE_COL_RESPONSABLE + " en " + hoja.Name);
+                            MessageBox.Show("No se ha encontrado la columna de responsable en " + hoja.Name);
                             continue; // si no se encuentra la columna que pase a la siguiente hoja
                         }//fin if no encuentra la columna Responsable
 
@@ -192,7 +73,7 @@ namespace NRFM_Auditoria
 
                             string nombre_responsable = hoja.Cell(colResponsable + index.ToString()).Value.ToString();
                             // se escribe el nombre del responsable en maysuculas y con un solo espacio para separar los nombres
-                            nombre_responsable = estandarizarNombre(nombre_responsable);
+                            nombre_responsable = FuncionesAuditoria.estandarizarNombre(nombre_responsable);
 
                             if (!xl_responsable.ContainsKey(nombre_responsable))
                             {
@@ -257,9 +138,9 @@ namespace NRFM_Auditoria
                             char colActual = 'A';
 
                             // buscamos la cabecera en la hoja actual 
-                            int Fin_Cabecera_HR = encontrarCabecera(xl_responsable[nombre_responsable].Worksheet(hoja.Name));
+                            int Fin_Cabecera_HR = FuncionesAuditoria.encontrarCabecera(xl_responsable[nombre_responsable].Worksheet(hoja.Name));
                             // obtenemos el fin de los datos de la hoja del responsable
-                            int finDatos = encontrarFinDatos(xl_responsable[nombre_responsable].Worksheet(hoja.Name), Fin_Cabecera_HR);
+                            int finDatos = FuncionesAuditoria.encontrarFinDatos(xl_responsable[nombre_responsable].Worksheet(hoja.Name), Fin_Cabecera_HR);
                             while (!hoja.Cell(colActual + Fin_Cabecera.ToString()).IsEmpty())
                             {
                                 // obtenemps el valor de la celda en el archivo de certificacion
